Renamed package to actors, added an example FingerTableService and an HashUtil class.
