+++ src/main/java/org/distributed/systems/chord/util/IHashUtil.java
@@ -1,0 +1,7 @@
+package org.distributed.systems.chord.util;
+
+public interface IHashUtil {
+
+    String hash(String input);
+
+}
