+++ src/main/java/org/distributed/systems/chord/util/impl/HashUtil.java
@@ -1,0 +1,11 @@
+package org.distributed.systems.chord.util.impl;
+
+import org.distributed.systems.chord.util.IHashUtil;
+
+public class HashUtil implements IHashUtil {
+
+    @Override
+    public String hash(String input) {
+        return null;
+    }
+}
